Folding angular_dynamic.dart into angular.dart
Also makes angular_static.dart export angular.dart

This allows a single 'angular.dart' import for the dynamic case which will be re-written from the transformer.
