+++ example/web/hello_world.dart
@@ -1,5 +1,4 @@
 import 'package:angular/angular.dart';
-import 'package:angular/angular_dynamic.dart';
 
 @NgController(
     selector: '[hello-world-controller]',
