+++ example/web/todo.dart
@@ -3,7 +3,6 @@ library todo;
 import 'dart:html';
 
 import 'package:angular/angular.dart';
-import 'package:angular/angular_dynamic.dart';
 import 'package:angular/playback/playback_http.dart';
 
 class Item {
