+++ lib/angular_core.dart
@@ -1,0 +1,13 @@
+library angular;
+
+
+export 'package:angular/bootstrap.dart';
+export 'package:angular/core/module.dart';
+export 'package:angular/directive/module.dart';
+export 'package:angular/core/annotation.dart';
+export 'package:angular/introspection.dart';
+export 'package:angular/filter/module.dart';
+export 'package:angular/routing/module.dart';
+export 'package:di/di.dart';
+export 'package:route_hierarchical/client.dart' hide childRoute;
+
