+++ test/bootstrap_spec.dart
@@ -1,7 +1,7 @@
 library bootstrap_spec;
 
 import '_specs.dart';
-import 'package:angular/angular_dynamic.dart';
+import 'package:angular/angular.dart';
 
 void main() {
   describe('bootstrap', () {
