+++ test/core/registry_spec.dart
@@ -1,7 +1,7 @@
 library registry_spec;
 
 import '../_specs.dart';
-import 'package:angular/angular_dynamic.dart';
+import 'package:angular/angular.dart';
 
 main() {
   describe('RegistryMap', () {
